(wip) All core tests are running with the new atom framework. Tests from other packages to follow.

+++ psl-core/src/main/java/org/linqs/psl/model/function/ExternalFunction.java
@@ -22,7 +22,7 @@ import org.linqs.psl.model.term.Constant;
 import org.linqs.psl.model.term.ConstantType;
 
 public interface ExternalFunction {
-    public double getValue(Database db, Constant... args);
+    public float getValue(Database db, Constant... args);
 
     public int getArity();
